move getNodeHints from NodeHelpers to frontend code
