+++ packages/workflow/test/NodeHelpers.test.ts
@@ -10,7 +10,6 @@ import {
 } from '@/Interfaces';
 import {
 	getNodeParameters,
-	getNodeHints,
 	isSubNodeType,
 	applyDeclarativeNodeOptionParameters,
 	getParameterIssues,
@@ -3459,95 +3458,6 @@ describe('NodeHelpers', () => {
 				expect(result).toEqual(testData.output.noneDisplayedTrue.defaultsTrue);
 			});
 		}
-	});
-
-	describe('getNodeHints', () => {
-		//TODO: Add more tests here when hints are added to some node types
-		test('should return node hints if present in node type', () => {
-			const testType = {
-				hints: [
-					{
-						message: 'TEST HINT',
-					},
-				],
-			} as INodeTypeDescription;
-
-			const workflow = {} as unknown as Workflow;
-
-			const node: INode = {
-				name: 'Test Node Hints',
-			} as INode;
-			const nodeType = testType;
-
-			const hints = getNodeHints(workflow, node, nodeType);
-
-			expect(hints).toHaveLength(1);
-			expect(hints[0].message).toEqual('TEST HINT');
-		});
-		test('should not include hint if displayCondition is false', () => {
-			const testType = {
-				hints: [
-					{
-						message: 'TEST HINT',
-						displayCondition: 'FALSE DISPLAY CONDITION EXPESSION',
-					},
-				],
-			} as INodeTypeDescription;
-
-			const workflow = {
-				expression: {
-					getSimpleParameterValue(
-						_node: string,
-						_parameter: string,
-						_mode: string,
-						_additionalData = {},
-					) {
-						return false;
-					},
-				},
-			} as unknown as Workflow;
-
-			const node: INode = {
-				name: 'Test Node Hints',
-			} as INode;
-			const nodeType = testType;
-
-			const hints = getNodeHints(workflow, node, nodeType);
-
-			expect(hints).toHaveLength(0);
-		});
-		test('should include hint if displayCondition is true', () => {
-			const testType = {
-				hints: [
-					{
-						message: 'TEST HINT',
-						displayCondition: 'TRUE DISPLAY CONDITION EXPESSION',
-					},
-				],
-			} as INodeTypeDescription;
-
-			const workflow = {
-				expression: {
-					getSimpleParameterValue(
-						_node: string,
-						_parameter: string,
-						_mode: string,
-						_additionalData = {},
-					) {
-						return true;
-					},
-				},
-			} as unknown as Workflow;
-
-			const node: INode = {
-				name: 'Test Node Hints',
-			} as INode;
-			const nodeType = testType;
-
-			const hints = getNodeHints(workflow, node, nodeType);
-
-			expect(hints).toHaveLength(1);
-		});
 	});
 
 	describe('isSubNodeType', () => {
